+++ site/dat/docs/Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 ## 1.3.0 (next)
  - add gatling script generation 
+ - fix grinder crash
 
 ## 1.2.0
  - maximize browser window in Selenium test, when possible